--- src/main/java/ru/reksoft/demo/config/Messages.java
@@ -1,35 +1,0 @@
-package ru.reksoft.demo.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
-import org.springframework.context.NoSuchMessageException;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.MessageSourceAccessor;
-
-import javax.annotation.PostConstruct;
-import java.util.Locale;
-
-@Configuration
-public class Messages {
-
-    private MessageSource messageSource;
-    private MessageSourceAccessor accessor;
-
-    @Autowired
-    public Messages(MessageSource messageSource) {
-        this.messageSource = messageSource;
-    }
-
-    @PostConstruct
-    private void init() {
-        accessor = new MessageSourceAccessor(messageSource, new Locale.Builder().setLanguage("ru").setScript("Cyrl").build());
-    }
-
-    public String get(String code) {
-        try {
-            return accessor.getMessage(code);
-        } catch (NoSuchMessageException e) {
-            return '{' + code + '}';
-        }
-    }
-}
