+++ src/main/java/ru/reksoft/demo/config/SecurityConfig.java
@@ -25,7 +25,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     }
 
     /**
-     * Differentiation of rights configuration
+     * Differentiation of rights configuration.
      */
     @Override
     protected void configure(HttpSecurity http) throws Exception {
