--- src/main/resources/ValidationMessages_ru.properties
@@ -1,1 +1,0 @@
-﻿
