+++ src/main/resources/messages.properties
@@ -1,3 +1,11 @@
-internal-server-error.message = Unexpected Internal Server Error. Please contact the administrator.
+#javax.validation.constraints.NotNull.message = must be set
+#javax.validation.constraints.NotEmpty.message = cannot be empty
+#javax.validation.constraints.Min.message = must be greater than or equal to {value}
+#javax.validation.constraints.Max.message = must be less than or equal to {value}
+#javax.validation.constraints.Size.message = size must be between {min} and {max}
 
-validation-error.message.title = Validation failed for %s
+reksoft.demo.validation.constraints.ReleaseYear.message = must be positive and not greater than the current year
+reksoft.demo.validation.constraints.PositionSequence.message = positions must follow the sequence
+
+reksoft.demo.validation.Throwable.message = Unexpected Internal Server Error. Please contact the administrator.
+reksoft.demo.validation.MethodArgumentNotValidException.message = Validation failed for %s.
