+++ src/main/resources/messages_ru.properties
@@ -1,3 +1,11 @@
-﻿internal-server-error.message = Непредвиденная ошибка. Пожалуйста, обратитесь к администратору.
+﻿javax.validation.constraints.NotNull.message = должно быть задано
+javax.validation.constraints.NotEmpty.message = не может быть пусто
+javax.validation.constraints.Min.message = должно быть больше или равно {value}
+javax.validation.constraints.Max.message = должно быть меньше или равно {value}
+javax.validation.constraints.Size.message = размер должен быть от {min} до {max}
 
-validation-error.message.title = Не удалось валидировать %s
+reksoft.demo.validation.constraints.ReleaseYear.message = должно быть положительным числом и не превышать текущий год
+reksoft.demo.validation.constraints.PositionSequence.message = позиции должны соблюдать последовательность
+
+reksoft.demo.validation.Throwable.message = Непредвиденная ошибка. Пожалуйста, обратитесь к администратору.
+reksoft.demo.validation.MethodArgumentNotValidException.message = Не удалось валидировать %s.
